refactor: configurable blog
