+++ .vscode/settings.json
@@ -1,5 +1,5 @@
 {
   "files.associations": {
     "*.mdx": "markdown"
-  }
+}
 }
