--- src/config/friends.ts
@@ -1,48 +1,0 @@
-import type { FriendCardProps } from "../components/Card/FriendCard/type";
-
-export const friends: FriendCardProps[] = [
-  {
-    name: "Airing",
-    link: "https://ursb.me",
-  },
-  {
-    name: "EMimTY",
-    link: "https://www.eaimty.com",
-  },
-  {
-    name: "Yuzi",
-    link: "https://yuzi.dev",
-  },
-  {
-    name: "RH-Xie",
-    link: "https://rhxie.top",
-  },
-  {
-    name: "7gugu",
-    link: "https://7gugu.com",
-  },
-  {
-    name: "songhn",
-    link: "https://songhn.com",
-  },
-  {
-    name: "younggglcy",
-    link: "https://www.younggglcy.com/"
-  },
-  {
-    name: "Talaxy",
-    link: "https://www.talaxy.site/"
-  },
-  {
-    name: "Lakphy",
-    link: "https://lakphy.me/"
-  },
-  {
-    name: 'cos',
-    link: 'https://ysx.cosine.ren/',
-  },
-  {
-    name: 'Corey Chiu',
-    link: 'https://coreychiu.com',
-  }
-];
