--- src/config/index.ts
@@ -1,5 +1,0 @@
-export const BLOG_TITLE = "Situ Note";
-
-export const BLOG_DESCRIPTION = "Personal website of situ2001";
-
-export const EMAIL = "dogecong@gmail.com";
