--- src/config/roles.ts
@@ -1,19 +1,0 @@
-export interface Role {
-  title: string;
-  place?: string;
-}
-
-const roles: Role[] = [
-  {
-    title: "Web Dev",
-    place: "TME",
-  },
-  {
-    title: "Open Source Contributor",
-  },
-  {
-    title: "Writer",
-  }
-]
-
-export default roles;
