--- src/config/site-info.mdx
@@ -1,6 +1,0 @@
-```toml
-# You can add me to your friends list with the following information
-title = "Situ Note"
-url = "https://situ2001.com"
-avatar = "https://situ2001.com/avatar.png"
-```
