+++ src/pages/rss.xml.ts
@@ -6,7 +6,7 @@ import {
   getPostsForRSS,
   posts as postsData,
 } from "../api/blog";
-import { BLOG_DESCRIPTION, BLOG_TITLE } from "../config";
+import config from "../../blog.config";
 
 const parser = new MarkdownIt();
 
@@ -14,8 +14,8 @@ export async function GET(context: any) {
   const posts = getPostsForRSS(postsData);
 
   return rss({
-    title: BLOG_TITLE,
-    description: BLOG_DESCRIPTION,
+    title: config.title,
+    description: config.description,
     site: context.site,
     customData: `<follow_challenge><feedId>69199567573101568</feedId><userId>41663616878351360</userId></follow_challenge>`,
     items: posts.map((post) => ({
@@ -28,7 +28,7 @@ export async function GET(context: any) {
         parser.render(
           "> 该内容使用MarkdownIt渲染，如需查看图片及获取更好的排版，请阅读原文\n" +
           "> This content is rendered using MarkdownIt, for better layout and images, please read the original post\n\n" +
-            post.body
+          post.body
         )
       ),
     })),
